app/course_service/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
PyJWT
aiokafka==0.8.1
